Add in-progress notes on the significance of some bin numbers to the game engine
These changes have been sitting on my disk for almost a year now.  While I do plan to return to this task, I don't expect to do so until some later date.  There is significant work here that I don't want to lose, so I am creating this temporary branch to hold the in-progress work (and any other incremental changes that I may trickle in) until this feature is complete and ready for inclusion in the default branch.