Update keyframe verb submode listing and PlayMode() info
The list of submode actions and numbers is now accurate. The previous list was
built from the comments in ky.pup, but testing has revealed that the submode
numbers in these comments are incorrect. The stye for the submode listing table
has been harmonized with the sound verb subclass listing table.