Fix navigation for tutorial pages on modern web browsers
Before this fix, the tutorial pages were not loading in the display frame due
to a cross-origin error. Modern browsers have more strict cross-origin policy
enforcement that prevents the interaction between frames that was previously
used to load the tutorial content. Namely, the browser blocks the chapter
selection frame from being able to read the location href of the display frame,
although it is still permitted to set it. The tutorial navigation mechanism has
been adjusted so it will function properly under these conditions.

One unresolved side effect of this change is that using the "Continue"
hyperlinks to advance through the tutorial pages no longer updates the chapter
selection to match the currently-displayed tutorial. However, all the tutorials
can be viewed despite this inconvenience.